+++ changelog.md
@@ -86,6 +86,7 @@ Codion Change Log
 - EntityTablePanel.editPanel() added.
 - EntityPanel control configuration refactored, now consistent with EntityEditPanel and EntityTablePanel.
 - EntityPanel, TabbedDetailLayout, now possible to configure the available panel states.
+- EntityComponents now final, constructor private, factory method added.
 
 ## 0.17.43
 ### is.codion.common.core